MysteryOfAntiques/requirements/MofA.txt
djangorestframework
django-rest-swagger